[#7557] [YCQL] Support != operator
Summary:
Currently, where/if clauses doesn't support the != operator. This is not an issue because these clauses support the NOT IN () operator and we can mimic semantics of != operator using NOT IN () with a single element in the list.

However, for language consistency and as syntactic sugar, we are adding != operator. This will supplement the existing = operator.

Also, a few points to note -

!= will only be allowed in select statements (both if and where clauses).
For WHERE clause -
i) no more than one usage of != operator will be allowed per column. The following operators can't be used with != operator: <,>,<=,>=,=,NOT IN (...), IN (...)
ii) != is not supported on subscripted columns (where the column is of jsonb/collection data type).
iii) != is, however, supported on JSONB/collection type columns when used without any subscripted args with following conditions -
a) JSONB type columns can have != with NULL/ JSON values.
b) Collections can only have != with NULL
For IF clause -
i) != operator can be used any number of times and in conjunction with any other condition. Point 2i) doesn't apply here.
ii) != is supported on subscripted columns.
iii) != is supported on unsubscripted JSONB/collection columns and both conditions a) and b) above still apply.

Test Plan:
ybd --scb --sj --java-test org.yb.cql.TestSelect#testNotEqualsQueryWithWhere
ybd --scb --sj --java-test org.yb.cql.TestSelect#testNotEqualsQueryWithIf

Reviewers: mihnea

Differential Revision: https://phabricator.dev.yugabyte.com/D10852