Currently, in an UPDATE statement - when a jsonb field with NULL value is set, the null value is always persisted to the underlying jsonb column.
This diff provides capability to skip updating fields in jsonb which are set to
NULL using a property in the WITH clause. This is needed for 2 use cases -
  1. Skip writing NULL to save some space
  2. If the field already has a non-NULL value in the jsonb column, skipping the
     NULL write would preserve existing value.

Usage:
  UPDATE table_name SET jcol->'field1' = NULL, jcol->'field2' = 2 WHERE
    pk = ? WITH options = {'ignore_null_jsonb_attributes' : true };

I copied some boiler plate code in pt_table_property.cc and pt_table_property.h
for pt_update_property.cc, pt_update_property.h. This was copied earlier as well
for pt_keyspace_property.cc. Separating out the common part into a super class
will be more code churn, time taking, and error prone. So avoiding for now to
fix this issue. Will separate it after we fix this.