[Backport 2.8] [#11234] YCQL: Fix flakiness in TestIndex.testIndexUpdateMisc and testPreparedStatement
Summary:
The flakiness in these tests was because the SELECT queries that were supposed
to use the index were executed before the index had reached the
INDEX_PERM_READ_WRITE_AND_DELETE permission as part of online schema change
during index creation. So, the queries were not using the index but the main
table and that gave a different result (which is a bug in itself -- check the
note below). The fix is to wait till the indexes reach the required permission
before executing the queries that we expect will use the index.

The changes in this diff are -
1. Use waitForReadPermsOnAllIndexes() in the two flaky tests before executing
   any queries that require using the index.
2. Use waitForReadPermsOnAllIndexes() in all other tests that are expecting the
   queries to use the index. This will make these tests robust since now we will
   surely be testing what we expect to test i.e., the index and not the main
   table.

NOTE: Presence or absense of waitForReadPermsOnAllIndexes() should not change
a query's result. Presence will merely ensure that we surely test the index for
the assertion and not the main table in the rare case when the index permission
hasn't reached INDEX_PERM_READ_WRITE_AND_DELETE till the query is executed. So,
the difference in result in testIndexUpdateMisc and testPreparedStatement is a
bug and should be fixed. That will be part of another diff after thorough
investigation since that is a day 0 bug.

Original commit: f1aca86, D15031

Test Plan:
Jenkins: rebase: 2.8
./yb_build.sh --java-test org.yb.cql.TestIndex

Reviewers: neil, mihnea, jason

Subscribers: yql, amitanand

Differential Revision: https://phabricator.dev.yugabyte.com/D15160