[Backport 2.12] [#11514] YSQL: Avoid raising spurious IllegalState for SELECT with LIMIT on kReadRestart
Summary:
The issue
---------
A SELECT statement with LIMIT clause prefetches data from tserver to reduce
latency in anticipation that some rows from existing rpc response might be
filtered out [see suppress_next_result_prefetching_ in PgDocOp::GetResult].

The prefetch rpc might face a kReadRestart error and then proceed further with
2 possible cases -

(1) In case the limit of rows isn't satisfied with the previous rpc's response,
the prefetch rpc's result is read and this results in propagation of the
kReadRestart error to the retry logic in postgres.c or the client.

(2) In case the limit of rows is satisfied already with the previous rpc's
response, the statement ends but the prefetch rpc marks the txn's read point for
read restart requirement. [read_point->RestartRequired in async_rpc.cc]

Later, for case 2, when the txn's commit is attempted, it leads to the following
illegal state - "Commit of transaction that requires restart is not allowed".

This illegal state is spurious for case 2 since the prefetch rpc's kReadRequest
error is not relevant -- since it is not used after the limit of rows is
satisfied by the previous rpc itself.

The fix
-------

For case 2, instead of raising an error on txn commit, we will just log a
warning. This is done by plumbing a has_used_prefetching_ field from
pg_doc_op.cc to transaction.cc and session.cc.

No tests are added to this diff since a later diff for READ COMMITTED has
test selectStarShortExecute_simpleQueryMode() that exposes this rare bug.

Original commit: https://phabricator.dev.yugabyte.com/D15549, TBD

Test Plan: Jenkins: urgent, rebase: 2.12

Reviewers: mihnea, mtakahara, sergei

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D15569