[#11630] [YSQL] Exception handling in stored procedure creates two internal sub-txns instead of one if READ COMMITTED isolation
Summary:
Exception handling in Pg stored procedures works by starting an internal sub
transaction at the start of the exception block to be able to rollback to it in
case an exception occurs. This is done by calling BeginInternalSubTransaction().

In READ COMMITTED isolation too, we start an internal sub transaction at the
start of each statement to be able to restart the statement. We do this in
StartTransactionCommand() by calling
BeginInternalSubTransactionForReadCommittedStatement() if the blockState is
TBLOCK_INPROGRESS, TBLOCK_IMPLICIT_INPROGRESS or TBLOCK_SUBINPROGRESS.

But BeginInternalSubTransaction() also calls StartTransactionCommand() and hence
results in an extra sub-txn on top of the one created for exception handling.
This messes with the exception handling since it expects the last added sub-txn
to be the one which is to be rolled back to.

The fix is to not call BeginInternalSubTransactionForReadCommittedStatement()
if StartTransactionCommand() was started by BeginInternalSubTransaction() for
exception handling.

The issue was caught by the yb_transactions_savepoints.sql test (run as part of
testPgRegressTransactionWithReadCommitted()) which has a case for
stored procedure with exception handling -
        "create function inverse(int) returns float8 as"

The test wasn't failing because testPgRegressTransactionWithReadCommitted() was
incorrectly setting "TEST_inject_sleep_before_applying_intents_ms=500" instead of
"yb_enable_read_committed_isolation=true".

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgRegressTransactionSavepoints#testPgRegressTransactionWithReadCommitted

Reviewers: mtakahara

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D15810