[WIP] [#12220] Use the correct value for ysql_output_buffer_size from relevant env variable
Summary:
ysql_output_buffer_size is a gflag that dictates how many bytes are stored in the
buffer from YSQL to the external client before requiring a flush. In READ
COMMITTED isolation, we guarantee that read restart errors will not be thrown
to the user as long as the output of a statement is within the
ysql_output_buffer_size limit (for reasons not mentioned in this summary).

ysql_output_buffer_size is read from FLAGS_ysql_output_buffer_size in pqcomm.c.
But if a user sets a value for gflag ysql_output_buffer_size on the tserver on
restart or cluster creation, FLAGS_ysql_output_buffer_size isn't updated on the
YSQL backend. Instead the env variable named "FLAGS_ysql_output_buffer_size" is
set in the YSQL processes to reflect the new value (this is as per
4d5bdc7c090caed3bb847fcdf34adb0508391f40).

So, to use the value set by the user on the tserver, pqcomm.c should read from
the env variable named "FLAGS_ysql_output_buffer_size" instead of directly
reading FLAGS_ysql_output_buffer_size.

Test Plan: TBD

Reviewers: mihnea

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D16635