[#12184] YSQL: Flush and wait for buffered ops before executing statements with non-transactional side-effects
Summary:
YSQL buffers tserver operations (writes specifically) unless it hits some
condition that forces it to flush the buffer and wait for the response. Some
conditions that force waiting for a buffer's response are - completion of a txn,
completion of an exception handling block in plpgsql, a read operation, or a
write to a key which already has a buffered write.

With buffering, execution can move on to later statements unless a flush and
response wait is required based on the conditions similar to those mentioned
above. For example - with autocommit mode off, writes for a statement (like
INSERT/UPDATE) are not flushed until required. Instead, they are buffered.
This is okay because, even before flushing, YSQL knows the number of
inserts/updates to be done and returns that number to the user client (as
"INSERT x"). If an error occurs in the rpc, it will anyway be caught in some
later flush before the txn commits (this was an optimzation introdcued in D7782).

Allowing execution to move on to later statements without waiting for the actual
work to be done on the tablet servers helps improve performance by buffering and
reduce latency. But, the downside is that any database violations associated with
the buffered operations might be observed after execution has moved to later
statements.

The downside is not an issue in many cases -- such as the optimization mentioned
above to not wait for writes of a DML in a transaction to complete. But, this can
cause issues with exception blocks as seen in gh issue #12184. Incorrect behaviour
is observed when an exception, that occurs due to some statement's rpcs, is seen
after some non-reversible side-effect(s) have also been executed. Non-reversible
side-effects are those which don't modify the database and hence can't be undone
when an exception is caught, or when a txn is rolled back.

E.g:
create table test (k int primary key, v int);
insert into test values (1, 1);
create function ...
returns table(z text)
begin
z := "before insert"
insert into test values (1, 1); -- operation is buffered, will result in duplicate key violation
return next; -- statement sends "before insert" to user. This is a non-transactional side effect of which can't be undone
exception
<some handler logic>
end;

The fix is: buffered operations should be flushed and waited for before
executing any statements with non-reversible side effects in plpgsql.

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgRegressBufferingInPlpgsql#testPgRegressBufferingInPlpgsql

Reviewers: dmitry, mtakahara

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D16809