[#11517] YSQL: Fix flakiness of TestPgTransparentRestarts.selectStarShortWithSavepoints
Summary:
TestPgTransparentRestarts.selectStarShortWithSavepoints is flaky because it
isn't seeing kReadRestart errors which it expects in a SELECT statement after
2 savepoint operations (the first operation is always a savepoint registration
but the second operation is either a savepoint rollback, release or another
registration). The test is performing 2 savepoint operations and expects
kReadRestart errors sometimes in the SELECT statement post that because retries
of kReadRestart errors by the query layer are avoided if savepoints have been
used.

However, the assertion that expects kReadRestart errors in some runs out of all,
fails. This is because of a prior SELECT, between the two savepoint operations,
which results in some local limits being picked for the tablet servers. This
reduces the chances of a kReadRestart in the second SELECT statement.

This diff fixes the flakiness by getting rid of the first SELECT. Also, the
SELECT statement had a LIMIT 1 clause. That is replaced by a SELECT count(*) to
increases the probability of hitting kReadRestart errors. The chances increase
because in the common case, LIMIT 1 would hit 1 tablet server and be done. The
count(*) would instead scan all tablet servers and increase the likely hood of
finding a recently committed kv pair that causes a kReadRestart error.

An extra case is added where the second operation is absent -- this tests the
scenario that kReadRestart errors are not retried by the query layer after the
first operation of savepoint registration.

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgTransparentRestarts#selectStarShortWithSavepoints

Reviewers: rsami

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D18232