[#12494] YSQL: Allow changing transaction characteristics in read committed isolation
Summary:
Currently in YSQL, a user isn't allowed to change the transaction isolation level
after starting a READ COMMITTED transaction. This is because of the following
reason:

(1) In PostgreSQL, the transaction isolation level can be changed only if no
sub-transactions have been created. If a sub-transaction is created the user
faces this error - "ERROR:  SET TRANSACTION ISOLATION LEVEL must not be called
in a subtransaction".

(2) In YSQL's read committed isolation, an internal saevpoint is created before
execution of each statement. This is an implementation detail and is required to
be able to undo work and restart a statement in case of serialization and
read restart errors. And each savepoint creates an internal sub-transaction. Hence,
when starting a transaction in READ COMMITTED isolation and trying to change
the isolation level using "SET TRANSACTION ISOLATION LEVEL ...", an internal
sub-transaction is started before changing the isolation level. This leads
to the above error. The same applies for "SET TRANSACTION DEFERRABLE" and
"SET TRANSACTION READ WRITE".

This diff fixes the issues by allowing change of transaction characteristics i.e.,
isolation, defferable, read-write if all sub-transactions are only internal
read committed sub-transactions and no statement/query was executed yet.

NOTE: The correct way to do this would have been to not start an internal
sub-transaction in the first place if a "SET TRANSACTION ..." statement is
encountered. However, that isn't possible without major code changes since
the internal sub-transaction is taken before parsing the statement.

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgIsolationRegress#isolationRegress

Reviewers: mtakahara

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D20688