[#14202] YSQL: Avoid rpc from YSQL to local tserver when switching sub-txns
Currently, YSQL sends the PgSetActiveSubTransactionRequestPB rpc to the local
tserver process' Pg client (pg_client_session.cc) when it wants to switch the
active sub transaction id. YSQL will switch the active sub transaction id for
any of following - user savepoint operations, an internal savepoint created for
exception handling or an internal savepoint created before execution of a
statement in READ COMMITTED isolation level as explained in 9f2cc7f.

However, a separate rpc is not required. The active sub transaction id can be
piggybacked in each PgPerformRequestPB through the PgPerformOptionsPB sub-field
and the Pg client session on the tserver can change the active sub transaction
id before handling the PgPerformRequestPB.

There are 2 benefits of avoiding the extra PgSetActiveSubTransactionRequestPB
rpc -

(1) Improved latency when creating sub transactions for a user savepoint or an
    exception block in PL/Pgsql procedures/functions.
(2) Lower latency for all statements in READ COMMITTED isolation because there
    is one less rpc at the start of each statement now for internal sub
    transaction creation.