[#14284, #13041] YSQL: Avoid returning read time in paging state when using serializable isolation level
Summary:
A regression occurred indirectly due 148d251e (which exposes the real issue). As
part of 148d251e, there was a minor refactor that rightly changed the meaning of
the `YBSession::SetReadPoint(const ReadHybridTime& read_time)` function in
session.cc. Before the change, the SetReadPoint(read_time) flavour of the
function would only set the non transactional read time. After the change, it
sets either the transactional or the non-transactional read time based on which
is applicable.

There were 4 instances of where SetReadPoint(read_time) was used -

(1) Two of those were called in a non-transactional context so the change in
SetReadPoint(read_time) didn't have any effect.

(2) One of those is called in YCQL's Executor::ExecuteAsync() which sets read
time if it is specified in "StatementParameters". And a read time is specified
in the statement params only if there was a paging state in the client request,
which in turn can occur only in a non-transational context (since YCQL doesn't
allow `SELECT` in a transaction).

The remaining instance of SetReadPoint(read_time) is in
`PgClientSession::SetupSession()` which is called when the read time is
specified in the "options" by YSQL. The change in meaning here causes the
regression due to the following reason:

(1) Before the change, if YSQL specified the read_time, it would be honoured (i.e.,
applied) only in a non-transactional setting in the `session`. After the change,
the read_time is honoured even when YSQL specifies it in a transaction.

(2) In a serializable transaction the tserver returns a read_time in the paging
state and YSQL sets the read time in the options as part of later PerformPB
rpcs. This leads to YSQL specifying the read time for later rpcs of the
serializable transaction, when it is not supposed to use any read time
(serializable transactions work on the latest data, not a snapshot of the
database).

So, the real issue is (2) which was exposed via (1). The fix is to not set the
read time in the paging state when using serializable isolation level.

Test Plan:
Jenkins: skip
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressPgMiscIndependent#testPgRegressPgMiscIndependent

Reviewers: dmitry, sergei

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D21071