[#18426] YCQL: Fix flakiness of TestTransaction.testReadRestarts
Summary:
This test run concurrent reads and writes until a desired number of kReadRestart
errors are seen. On mac, it takes much longer than the test timeout for those
many errors to occur. This change modifies the test to run the workload for a
maximum of 100 seconds. If the desired number of errors occur are reached
earlier (which happens in debug/ release builds), the workload is stopped.

Note that asan and tsan builds also take a long time to reach the
desired number of errors but they haven't been failing since they have a higher
test timeout. Going forward, these tests will also only run for a maximum of
100 seconds.

Test Plan:
./yb_build.sh release --java-test org.yb.cql.TestTransaction#testReadRestarts -n 10 --tp 1
Jenkins: test regex: .*testReadRestart.*

Reviewers: smishra

Subscribers: yql