[#27037] YSQL: Allow concurrent DDLs when object locking is enabled
In YSQL, we restrict successful execution of concurrent DDLs on the same database
because we don't have a solid serialization mechanism in place to detect
and allow DDLs which can be safely executed concurrently. There are some exceptions
to this such as concurrent CREATE TABLEs.

To restrict concurrent DDLs, we rely on a 2-prong mechanism:

(1) Use snapshot isolation to detect conflicting writes from concurrent DDL transactions.
(2) Almost all DDLs increment the per-db catalog version at the end of the DDL transaction.

This ensures that any 2 concurrent DDLs that update the catalog version face a
write-write conflict on the catalog version. Note that two DDLs conflict because
they have an overlapping `[read time, commit time]` window and in Snapshot
Isolation we can't have two writes to the same piece of data in this overlapping
window.

For DDLs such as `CREATE TABLE`, we by-pass this by not incrementing the catalog version.

Postgres acquires object locks for serialization. Concurrent DDLs that operate on the
same object in a conflicting manner wait on each other. And once a DDL is unblocked
and acquires a lock, it process invalidation messages to ensure that it works based on the
latest catalog information. So, the cache is always kept coherent. (see AcceptInvalidationMessages() in
lmgr.c).

YSQL recently added support for object locking under the gflag
`enable_object_locking_for_table_locks`. With object locking in place,
we can serialize concurrent DDLs similar to Pg.

This diff gets rid of the YSQL mechanism to restrict concurrent DDLs if
the gflag `enable_object_locking_for_table_locks` is set. For this, we will
reset the hybrid time for each read/ write rpc of a DDL statement. This results
in picking the current time as the read time for each rpc after acquiring
in-memory conflict checking locks.

Conflict detection check for conflicts in regular and intents db. Using the current time helps avoid kConflict errors as follows:

(1) No conflict will be found with committed data in regular db because the read time will
be higher than all committed data.

(2) No conflict will be found in intents db because object locks will ensure that we don't see concurrent transactions that have provisional writes to the same piece of data.

Note that this is not exactly Pg implementation -- Pg uses a catalog snapshot that
is invalidated after processing invalidation messages. We will eventually match
Pg's implementation as well by integrating Pg's catalog snapshot logic
with YSQL read points (similar to 9124c053).

Since we will be using the latest time as the read time always, for serializing writes to
the catalog version and invalidation messages table, we will acquire exclusive object
locks on both tables. We still want to keep the catalog version increments since the
version is used as a key to the invalidation messages table.

NOTE: Even though all read/ write rpcs in DDLs use the latest time as the read time,
reads done as part of the kCatalog kind in PgClientSession will still use an old read time and return stale data. This can lead to issues such as https://github.com/yugabyte/yugabyte-db/issues/28493. This will be fixed in a
follow-up diff.