[#29362] YSQL: Add GUC to allow toggling of FK correctness fix
Summary:
As part of f225ca435, fk triggers started using a fresh snapshot instead
of the transaction snapshot. This was required to fix correctness issue
described in that commit. For this fix, a prequisite commit (18eb1a9fc5) was
required to integrate PG's snapshot managament with the analogous
read_time_serial_no management in pggate. This integration was already in place
for READ COMMITTED isolation, the prereq commit extended it to
REPEATABLE READ isolation level.

This revision introduces a GUC to allow a user to fallback to the mode
before both commits. Use of the GUC is not recommended and needs extreme
caution.

NOTE: this is a 2025.1 only commit. The fallback option is not added to
future releases to avoid complexity in code that is actively bieng
updated. Also, future releases will have much more soak time for the
changes.

Test Plan: Jenkins: rebase: 2025.1

Reviewers: smishra

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D48305